Chore: Separating derived classes of input validator
